Dzejkop/reth-1.1.4 (#79)
* add update OP Payload Builder type

* wip

* add world chain builder

* wip, fix imports

* fix witness

* temp remove tests

* Massive reorg (reflects optimism directory within reth)

* Fix primitives deps

* Remove old bin

* wip

* wip

* wip

* Back to square 1

* @0xKitsune's changes

* fmt

* fix

* Code compiles - tests don't

* Everything compiles

---------
