+++ world-chain-builder/Cargo.toml
@@ -1,14 +1,38 @@
-[package]
-name = "world-chain-builder"
+[workspace.package]
 version = "0.1.0"
 edition = "2021"
-
-default-run = "world-chain-builder"
+rust-version = "1.82"
+license = "MIT"
+homepage = "https://world.org/world-chain"
+repository = "https://github.com/worldcoin/world-chain/"
 
 [workspace]
-members = ["crates/*"]
+resolver = "2"
+members = [
+    "crates/assertor",
+    "crates/toolkit",
+    "crates/world/node",
+    "crates/world/bin",
+    "crates/world/payload",
+    "crates/world/rpc",
+    "crates/world/pbh",
+    "crates/world/db",
+    "crates/world/primitives",
+    "crates/world/pool",
+]
 
-[dependencies]
+[workspace.lints]
+
+[workspace.dependencies]
+# internal
+world-chain-builder-node = { path = "crates/world/node" }
+world-chain-builder-db = { path = "crates/world/db" }
+world-chain-builder-pbh = { path = "crates/world/pbh" }
+world-chain-builder-payload = { path = "crates/world/payload" }
+world-chain-builder-primitives = { path = "crates/world/primitives" }
+world-chain-builder-rpc = { path = "crates/world/rpc" }
+world-chain-builder-pool = { path = "crates/world/pool" }
+
 # reth
 reth = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
 reth-cli-util = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
@@ -19,12 +43,21 @@ reth-provider = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4", f
     "test-utils",
 ] }
 reth-basic-payload-builder = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
+reth-transaction-pool = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
+reth-rpc-server-types = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
 reth-prune-types = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
 reth-trie = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
 reth-chain-state = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
 reth-eth-wire-types = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
 reth-stages-types = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
 reth-trie-db = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
+reth-primitives-traits = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
+reth-node-builder = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
+reth-payload-builder = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
+reth-payload-util = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
+reth-payload-validator = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
+reth-revm = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4", default-features = false }
+reth-tasks = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
 
 # reth-optimism
 reth-primitives = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4", features = [
@@ -49,11 +82,19 @@ reth-optimism-chainspec = { git = "https://github.com/paradigmxyz/reth", tag = "
 reth-optimism-payload-builder = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
 reth-optimism-forks = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
 reth-optimism-primitives = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
+reth-e2e-test-utils = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
+reth-consensus = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
+reth-node-core = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
+reth-tracing = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
 
-# alloy
+
+# alloy op
 op-alloy-consensus = "0.7.3"
 op-alloy-rpc-types = "0.7.3"
+op-alloy-rpc-types-engine = "0.7.3"
 op-alloy-network = "0.7.3"
+
+# alloy
 alloy-consensus = { version = "0.7.3", default-features = false }
 alloy-network = { version = "0.7.3", default-features = false }
 alloy-primitives = { version = "0.8.11", default-features = false }
@@ -61,28 +102,25 @@ alloy-rpc-types-eth = { version = "0.7.3", default-features = false }
 alloy-rpc-types = { version = "0.7.3", features = [
     "eth",
 ], default-features = false }
+alloy-rpc-types-engine = { version = "0.7.3" }
 alloy-rlp = "0.3.10"
 alloy-eips = { version = "0.7.3", default-features = false }
 alloy-genesis = { version = "0.7.3", default-features = false }
+alloy-rpc-types-debug = "0.8.0"
+alloy-signer = { version = "0.7.3", default-features = false }
+alloy-signer-local = { version = "0.7.3", default-features = false }
 
 # revm
 revm = { version = "18.0.0", features = ["std"], default-features = false }
-revm-inspectors = "0.12.0"
-revm-primitives = { version = "14.0.0", features = [
-    "std",
-], default-features = false }
-revm-interpreter = "10"
-revm-precompile = "11"
-
+revm-primitives = { version = "14.0.0", default-features = false }
 
 # 3rd party
-jsonrpsee = { version = "0.24", features = ["server", "macros"] }
+jsonrpsee = { version = "0.24", features = ["server", "client", "macros"] }
 tokio = { version = "1", features = ["full"] }
 futures = "0.3"
 chrono = "0.4"
 thiserror = "1"
-strum = "0.26"
-strum_macros = "0.26"
+strum = { version = "0.26", features = ["derive"] }
 bytemuck = "1"
 semaphore = { git = "https://github.com/worldcoin/semaphore-rs", rev = "d0d1f899add7116ccc1228f5e5e5ee2e2e233768", features = [
     "depth_30",
@@ -94,34 +132,14 @@ tracing = "0.1"
 parking_lot = "0.12"
 derive_more = "1"
 dotenvy = "0.15.7"
-tikv-jemallocator = { version = "0.6.0", optional = true }
+tikv-jemallocator = { version = "0.6.0" }
 bytes = "1.7.2"
 hex = "0.4.3"
 tempfile = "3"
-
-[dev-dependencies]
 criterion = { version = "0.5", features = ["async_tokio"] }
 serial_test = "3"
 test-case = "3"
 ethers-core = { git = "https://github.com/gakonst/ethers-rs", default-features = false }
-alloy-primitives = "0.8"
 serde_json = "1"
 rand = "0.8"
-reth-e2e-test-utils = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
-reth-consensus = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
-reth-optimism-consensus = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
-reth-node-core = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
-reth-tracing = { git = "https://github.com/paradigmxyz/reth", tag = "v1.1.4" }
-alloy-signer = { version = "0.7.3", default-features = false }
-alloy-signer-local = { version = "0.7.3", default-features = false }
-
-[features]
-jemalloc = ["tikv-jemallocator"]
-
-[[bin]]
-name = "world-chain-builder"
-path = "bin/world-chain-builder.rs"
-
-[[bench]]
-name = "validate_transaction"
-harness = false
+reqwest = { version = "0.12", default-features = false }
