+++ world-chain-builder/crates/world/db/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "world-chain-builder-db"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+reth-db = { workspace = true }
+reth-db-api = { workspace = true }
+reth-primitives = { workspace = true, features = ["optimism"] }
+revm-primitives.workspace = true
+
+semaphore = { workspace = true }
+serde = { workspace = true }
+tracing = { workspace = true }
+bytes = { workspace = true }
+eyre = { workspace = true }
