+++ world-chain-builder/crates/world/node/Cargo.toml
@@ -1,0 +1,50 @@
+[package]
+name = "world-chain-builder-node"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+world-chain-builder-db.workspace = true
+world-chain-builder-payload.workspace = true
+world-chain-builder-pool.workspace = true
+
+reth.workspace = true
+reth-chain-state.workspace = true
+reth-db.workspace = true
+reth-optimism-chainspec.workspace = true
+reth-optimism-node.workspace = true
+reth-optimism-primitives.workspace = true
+reth-basic-payload-builder.workspace = true
+reth-optimism-payload-builder.workspace = true
+reth-primitives.workspace = true
+reth-primitives-traits.workspace = true
+reth-provider.workspace = true
+reth-prune-types.workspace = true
+reth-trie.workspace = true
+reth-trie-db.workspace = true
+
+alloy-eips.workspace = true
+alloy-primitives.workspace = true
+alloy-rpc-types.workspace = true
+
+tokio.workspace = true
+eyre.workspace = true
+futures.workspace = true
+clap.workspace = true
+
+[dev-dependencies]
+criterion.workspace = true
+semaphore.workspace = true
+
+world-chain-builder-pool = { workspace = true, features = ["test"] }
+
+[[bench]]
+name = "validate_transaction"
+harness = false
