+++ world-chain-builder/crates/world/node/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod args;
+pub mod node;
+
+// TODO: Conditional compilation?
+// #[cfg(test)]
+pub mod test_utils;
