+++ world-chain-builder/crates/world/node/test/assets/genesis.json
