+++ world-chain-builder/crates/world/payload/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "world-chain-builder-payload"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+# world-chain
+world-chain-builder-pool.workspace = true
+world-chain-builder-rpc.workspace = true
+
+# reth
+reth.workspace = true
+reth-basic-payload-builder.workspace = true
+reth-chain-state.workspace = true
+reth-evm.workspace = true
+reth-db.workspace = true
+reth-optimism-node.workspace = true
+reth-optimism-chainspec.workspace = true
+reth-optimism-consensus.workspace = true
+reth-optimism-payload-builder.workspace = true
+reth-provider.workspace = true
+reth-primitives.workspace = true
+reth-transaction-pool.workspace = true
+reth-trie.workspace = true
+
+# revm
+revm.workspace = true
+revm-primitives.workspace = true
+
+# alloy
+alloy-consensus.workspace = true
+alloy-eips.workspace = true
+alloy-rpc-types-debug.workspace = true
+
+# misc
+tracing.workspace = true
+jsonrpsee.workspace = true
