+++ world-chain-builder/crates/world/pbh/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "world-chain-builder-pbh"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+alloy-rlp.workspace = true
+
+chrono = { workspace = true }
+thiserror = { workspace = true }
+semaphore.workspace = true
+strum.workspace = true
+serde.workspace = true
+
+[dev-dependencies]
+ethers-core.workspace = true
+test-case.workspace = true
