+++ world-chain-builder/crates/world/primitives/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "world-chain-builder-primitives"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+world-chain-builder-pbh.workspace = true
+
+reth.workspace = true
+reth-primitives = { workspace = true, features = ["optimism"] }
+
+revm-primitives.workspace = true
+
+alloy-eips.workspace = true
+alloy-rlp.workspace = true
+
+tracing.workspace = true
+bytes.workspace = true
+
+[dev-dependencies]
+alloy-consensus.workspace = true
+alloy-primitives.workspace = true
