+++ world-chain-builder/crates/world/rpc/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "world-chain-builder-rpc"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+world-chain-builder-pool.workspace = true
+
+reth.workspace = true
+
+reth-provider.workspace = true
+revm-primitives.workspace = true
+alloy-consensus.workspace = true
+
+alloy-eips.workspace = true
+alloy-primitives.workspace = true
+alloy-rpc-types.workspace = true
+
+jsonrpsee.workspace = true
