--- world-chain-builder/src/lib.rs
@@ -1,7 +1,0 @@
-pub mod node;
-pub mod payload;
-pub mod pbh;
-pub mod pool;
-pub mod primitives;
-pub mod rpc;
-pub mod test;
