--- world-chain-builder/src/node/mod.rs
@@ -1,5 +1,0 @@
-pub mod args;
-pub mod builder;
-
-#[cfg(test)]
-pub mod test_utils;
