--- world-chain-builder/src/pbh/mod.rs
@@ -1,4 +1,0 @@
-pub mod date_marker;
-pub mod db;
-pub mod external_nullifier;
-pub mod payload;
