--- world-chain-builder/src/pool/mod.rs
@@ -1,7 +1,0 @@
-pub mod builder;
-pub mod error;
-pub mod noop;
-pub mod ordering;
-pub mod root;
-pub mod tx;
-pub mod validator;
